Refactor two_factor_auth into separate method
This commit puts the two factor code into it's own method, and checks
if two factor is needed from both the send_request and the
login methods. Therefore, even if feedback_required is returned from the
send_request method, the second check in the login method will still
activate 2FA if needed.