Replace spaces in code
As I found out, if you enter the code just as you receive it via SMS (e.g. "123 123") the code is not accepted. You have to enter the code without spaces. To make it more userfriendly in case some users enter the code just as they receive it, i remove spaces from the input.